tax-sale-agent/requirements.txt
pandas==2.2.2
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
pdfplumber==0.11.4
tabulate==0.9.0
playwright==1.47.2
